refactor background services
